fix(TaskSnapshot): change Version type from int to long for better compatibility

+++ StateStore/Models/TaskSnapshot.cs
@@ -3,9 +3,9 @@ namespace JojaAutoTasks.StateStore.Models;
 internal sealed class TaskSnapshot
 {
     internal IReadOnlyList<TaskView> TaskViews { get; }
-    internal int Version { get; }
+    internal long Version { get; }
 
-    internal TaskSnapshot(IReadOnlyList<TaskView> taskViews, int version)
+    internal TaskSnapshot(IReadOnlyList<TaskView> taskViews, long version)
     {
         TaskViews = taskViews;
         Version = version;
